The Timer ISR trigger a stack fault (0x0C)
The leave instruction in the Timer ISR triggerred a  stack fault.
This instruction is used by C in an C-function and I tried to
use a C funciton as an interrupt handler, so I mimic it.
However, it seems that using C as an interrupt handler is not
feasible.

Except that, things are going find.