Without stack fault. Timer not well programmed.
The stack fault is because the assumeption of my asm code on the stack
is not identical with that of the C compiler.

I didn't assume that a block of stack was used for the argument of the
function call. While this block was pushed the %ebp into, it's content was
corrupted when used for the argument. The value of %ebp was not right
after pop-ed. This ill %ebp value cause a stack fault in the following
instruction.

The ISR is now triggerred onece and returned successfully.

The Timer device (8253) should be correctly programmed to generate
continuous interrupts.