Timer interrupt full function.
Because I forgot to 'recogized' the acception of the interrupt to 8259,
that is, to send 0x20 command to port 0x20/0xA0, the 8259 suspended
the following interrupt. This is why Timer ISR was executed just once.

The ISR is still coded manually in assembly. How to coded them neatly?

Program architecture is kind of not so good.