Loaded TR
A misunderstanding cost me hours.
TSS descriptor should not be set busy in the beginning.
The TR instruction is going to set it to busy.

At last, we need not to initialize TSS at all because the loaded TSS is
recognized as the active one, the content is going to be maintained by
processor (almost, except for the IO map etc.)