Stop sending characters in user mode.
The segment selector to screen buffer is of privilege level 0. Using
this selector should yield protection fault.

But Bochs does not trap this fault. This did caused a problem in qemu
and virtual box, but I cannot explain the stack of qemu at entering int
0x0d. The 'return address' in the stack points to a place that is
unlikely to violate protection. It just mov a const into %eax.

But because %ds was always 0x20 when protection fault, it is most
suspicious that the Screen::putchar() is the root cause. And the
phenomenon disappeared after stopping writing to screen buffer in user
mode.

Considering linking i386-stub.c to hOS. Becuase bochs can not send out
serial port content as expected. We might need to debug on qemu or
virtual box. And we are going to debug on a physical machine finally.